requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
Augmentor==0.2.10
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
fonttools==4.38.0
future==0.18.3
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==2.10.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
ipykernel==6.15.2
ipython==7.31.1
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==4.11.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.18.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.4.1
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==2.2.0
tifffile==2021.11.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.10
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.13.0
